Build manifest from paths in parallel
